Configure TravisCI to send notifications to BitBot
This entry in the .travis.yml file forwards a verbose payload of information
to our lita travis handler for all travis ci notifications.  Since Travis
does not provide a way to filter notifications by branch in the .travis.yml
file, we will do the filtering and formatting in the lita handler.