Use custom fork of stripes-core.
We need to make changes to core stripes libraries from time to time,
and we don't necessarily want to have to wait on them to get
merged. The branch `eholdings-demo-support` on a given library will
contain whatever set of independent patches are necessary to make the
eholdings demo platform and test cases work.

This allows the individual enhancements to not be dependent on each
other and submitted and considered as separate PRS. There is currently
only one such enhancement, but there could be many.

As PRS make it to master, they should be removed from this branch.