Don't roll a separate package for demo app.
The only thing preventing us from running our tests and a demo
application was that you need to be able to require the package under
development `@folio/eholdings` from within the test harnes and the
demo code. We were accomplishing this by having a separate `platform/`
directory with its own `package.json` and `node_modules/` so that it
could load the eholdings module.

This sidesteps the issue entirely by adding a `postinstall` script
called to symlink the eholdings packages module into its own
`node_modules` directory. Now any node process that runs from the root
of the eholdings package directory will be able to load the eholdings
module.

By only having a single `node_modules` and a single directory to
contend with, this drastically simplifies the development workflow and
reduces the need for re-installs.