Select / deselect package titles (customer resources)
One of the primary use-cases for the e-holdings application is for
digital resource librarians to manage their holdings. To do this they
need the ability to mark a title as being part of their
holdings (toggling it on and off);

This work, along with recent changes to stripes-core, leverages the
concept of 'epics' from 'redux-observable' to ensure async actions
resolve sanely.