Select / deselect package titles (customer resources)
One of the primary use-cases for the e-holdings application is for
digital resource librarians to manage their holdings. To do this they
need the ability to mark a title as being part of their
holdings (toggling it on and off);

This adds a reducer to the CustomerResourceShowRoute which handles all
of the state changes. Currently, all of the side-effects are handled
in `mapDispatchToProps`, but we should consider moving this over to
using `redux-observable`.