Ensure reliable click events in customer resource test
We were seeing flakiness around the tests related to selecting/deselcting packages.
These enhancements ensure the existence of an element to be clicked before performing the click.
We also temporarily increase the mirage server timing to account for requests completing before the convergent `it` gets a chance to assert anything about in-flight requests.