Opt in to Mirage.
Up until this point, we've been using Mirage as the default backend
for running our development server. This has served us well as we
built up the API, but nowadays, we will go large spans of time where
we aren't changing the backend at all, but only the UI that is built
on top of it.

It makes sense then to make running against the (somewhat) stable
backend as the default. However, if you want to enable Mirage then you
can do so by setting the `MIRAGE` environment variable