Checkin yarn lock.
While this is a library, and not an app, the actual demo app that's
being shipped is what we're using in production.

If we need to test against multiple dependencies, then we use a
mechanism like *-try that does multiple installs with diffent version.

In either case though, having deployment stability is critical.