View and set backend settings
As it stood, the value for the RMAPI customer ID and API key were per
deployment. Now however, it can be configured via the
/eholdings/configuration endpoint so that each tenant can maintain
their own set of credentials. We needed a way to detect if the
eholdings app is usable, and if its backend is configured
correctly, and also to be able to change credentials.

This change checks on every eholdings application entry to see if
infact there _is_ a  backend for eholdings and gives you instruction
on how to download it. If there is a backend, it will ensure that it
is configured properly, and if not, it will direct the user to a new
settings screen with which a tenant admin can view the current
configuration and make changes where necessary.

In order to support development against server configurations which
are hard to arrive at (such as a borked backend) the ability to start
the app in different scenarios was added. For example to start the app
without a backend, you would invoke the command:

```
yarn start --mirage no-backend
```