Fix triggering a search when switching search types
Previously, in a narrow view when switching search types, it would
trigger the search pane to close due to the presence of a new `search`
query in the URL. Additionally, this trigger caused a new search to be
made with the old query but within a different search type.

Now in both `componentWillReceiveProps` hooks, we check if we are
using the same search type as before, before creating a new search or
closing the open search pane.

A few tests were added to make sure switching between search types
keeps the old results when switching back (and doesn't actually cause
a new query in a different search type context).