Add preview pane to search
The main child routes for each resource is now nested under the search
route. This search route detects the presence of the `searchType`
query parameter and will render the search form and results. When the
`searchType` parameter is missing, it will instead just render it's
children (which are the various show routes).

The initial load of any of the search routes will now perform an empty
search, returning the first page of all results from the API.