Remove auto search when changing search types
When this new search happens via `componentWillReceiveProps` it seems
as though the `connect` function does not receive the updated state
during the request actions.

When a request is fired this way, a re-render does happen with every
action, however the search leaf returned from the `connect` function
remains the same. Once the request is complete, switching to another
search type and back will trigger the new search state to be rendered.

For now, this auto-search feature is removed so we can push out the
three-pane search layout