Create impagination and query components
The Impagination component is responsible for paging and handling the
impagination dataset and providing the records of that dataset state.

The Query component is responsible for keeping track of model
collections by which page the collection is associated with. The Query
component will intelligently call the `fetch` property when a paged
collection is requested from the Impagination component it renders.

The various `*-search-results` components have been replaced with
`*-search` components to indicate that these component are now
responsible for deciding when to make a query request. The route will
still provide the actual request action, but the component will call
this action as necessary.