Add scroll tracking to query component and rename to query-list
The QueryList name better describes that the component is also
responsible for rendering a list of items. The Impagination component
has also been moved to be beside the QueryList component as that's
the only time the Impagination component is used (and the two
components were also designed to work in conjunction with each other).

The QueryList component now has the ability to track the scroll
position of the list it renders. It continually sets the read offset
of the impagination component and calls `fetch` when impagination
requests new data. `onPage` will be called when the page threshold has
been crossed and a new page is to be considered "in view".