Implement scroll tracking requirements for search lists
With these requirements in place, the search route can now update
the page query param when the `onPage` prop of the query-lists
are triggered.

In order for the query-list component to properly track the scroll
position of a list, each list element must have a predictable
height. This height is provided to the query-list via `itemHeight`.

The various list-item components were given styles to ensure their
sizes are predictable. In addition to their height, the names will
now overflow with ellipsis when they are too long to fit onto a
single line.

The *-search components have been renamed to *-search-list to better
reflect that they render a query-list (which renders a list).