Add error handling to query-list component
The query-list component will now render any errors it encounters for
a record in the list. If there are no records, it will use the first
rejected page's error message. If there are no errors, and no records,
it will render a `notFoundMessage`.

The testing page objects have been updated to reflect the new
query-list data-test-attributes.