Fix relationships when not using include param
If there is no relationship data, but there is a relationship request,
get the related records via that newer request.

One of the changes to mod-kb-ebsco was to return an updated payload in
response to PUT requests. These updated payloads did not include
relationships and so related properties were not properly finding the
correct associated records.

In mirage, when you return a record from a PUT request it does not
include any information about the relationships. So the reducer kept
the old relationships in that case; enabling the resolver to retain
knowledge about which records to return. The reducer has now been
updated to clear out any old attributes or relationships when it is
updating a record in the data store. This update to the reducer now
allows us to use mirage to test what is actually happening in our
production environment.

Finally, there was a false positive in our tests for all titles in a
package being selected. The test did not check the length of titles,
so when no titles are displayed, they were all "selected" ([].every
returns true). This test has been adjusted to return false when there
are no titles being displayed.