Move pagination into collection class
A collection now manages all of the pages for a request based on the
other params provided to the collection.

This allows us to remove the collection tracking from the query-list
component and go back to only providing collections via route components.

This also reimplements the results total for the paneset component
since we have access to the total collection length in the route.