Fix false positive tests
Because of the speed of the tests, when this "second" search was made
the first search never actually happened. Converging on the first
search's success before making the second search surfaces an issue
with making new searches.

The fix was to check every componentWillReceiveProps hook between the
search route and the query-list component; making adjustments where
necessary to ensure the proper collection was rendered.

Even after making the adjustments, the impagination component needed to
create a new dataset when the search changed. Since new collection
instances are created all of the time, we need to check if it is a
newly loading collection before creating the new dataset for it.