Break away scroll view from query-list component
This splits the new components up and creates an additional one.

The impagination component is responsible for resolving records from a
collection and fetching additional records when needed.

The scroll-view component is responsible for rendering and scrolling
through a list of items and only rendering the items that should be
visible in the viewport.

The query-list component combines the impagination and scroll-view
components to create an infinite-scroll paging-list for a collection
of records. This component also renders any errors related to the
provided collection, and could potentially show a loading indicator
for each item in the list.