Use offset instead of page in the search URL
Also move totalPages into the collection, since it reflects some state
of the collection.

A key was added to collections to better help differentiate between
them. Since a new collection is created whenever the data store state
updates, it is difficult to determine if a collection is truly new, or
simply an update to an previously existing collection. The key
property is unique to a set of query params for a resource type,
regardless of the page param. Different keys will belong to different
requests for records.