Debounce offset and support empty arrays in scroll-view component
Debouncing the offset update helps to prevent uneccessary calls to the
pages between scroll positions when scrolling a long list of items.

Sometimes, slicing the dataset state produces an empty array. We now
patch the dataset state and introduce a slice method which uses
`getRecord` to always return an array of records.