Use overflow: auto
`overflow: scroll` causes the scrollbar to always be visible even with
no scrollable content. When there is nested scrollable content, this
causes a double scrollbar.

This isn't noticable on an OS that hides scrollbars, but viewing on
another OS, or setting the scrollbars to always be visible surfaces
this issue.