Hide scrollbars for list-item elements
Previously, the overflow-y was allowed to be visible due to descenders
being cut off. The line-height has since been adjusted and the
overflow-y property is causing permanent scrollbars to be visible for
these elements. Since the line-height is adjusted, we can safely hide
the overflow on these elements without cutting off descenders.