Check for foreign key attributes in `belongsTo` relationships
The internals of the `belongsTo` method for model classes has been
augmented to check for possible foreign keys (i.e. packageId,
vendorId, etc.) in a record's data attributes, in addition to the
check it was already doing on the explicit `relationships` key.

This allows for more reliable linking of related records already
in the store, and eliminates some UI bugs that were found to be a
result of incomplete linking functionality.  Now models can find their
related models' data without an explicit use of JSONAPI's `include` parameter.