Don't link self when pulled in as a dependency.
When ui-eholdings is being brought in as a dependency of a stripes
platfom build, and is not the platform itself, the node_modules are
flattened out into the module that contains eholdings as a
dependency. This means that the `bin/link-self` script fails because
it expects node_modules to exist already, but that won't happen until
yarn links dependencies from the root directory.

This is a bit of a hack, but we have the `link-self` script check for
the availability of the node_modules directory _first_, and only then
proceed if it exists.

This is acceptable since the entire `bin/link-self` script will go
away completely once UI eholdings can be built on stripes-cli.