Make inflected a primary dependency.
When using ui-eholdings as a dependency, it doesn't install its
devDependencies. This isn't a problem when running the CI build
because it's installed as the root package so all the dev-dependencies
make it in, but when running it as part of a containing stripes
platform, it breaks because it doesn't bring in things it needs to
build.

This adds:

- inflected
- query-string

as hard dependencies of ui-eholdings so that it will always get pulled
into every build of which it is a part.