Fix a typo while clearing out customizations during deselection
We noticed an issue when deselecting customer resources, and traced it
to an issue with embargo values.  Looks like we had a typo in there,
and were clearing out the wrong field during deselection.  This work
adjusts the clear-out logic to use empty objects where possible, in
lieu of nulling out specific data members within those objects.