Optimize resolver request matching
Added a filter option for `resolver.getRequest()` so that the request
getter for collections do not have to loop over a theoretical amount
of pages. Instead, it will now just find the first matching resolved
request in the store as oppose to searching the store for a request
for every page.

Also fixed a bug when matching the `path` of requests where package
customer-resource requests were being returned for title
customer-resources because there is no path for title
customer-resources. When `path` was `undefined`, the request matcher
didn't bother matching on the path. Now it will still attempt to match
on the path as long as `path` key exists.