Rewrite impagination component to always update records
When records are resolved, impagination keeps a reference to the
record in our store. When the record is updated immutably, the old
record becomes stale and is not reflected in UI using the impagination
component.

By using impagination's state machine directly we can resolve or
reject records at will without promises. This allows us to update any
records currently in the load horizon with accurate records from our
own data store.