Make PR template more user friendly.
Our PR template doesn't provide support to folks new to the project,
and is instead more geared towards people who have already "drunk the
koolaid" so to speak. It provides terse boilerplate that can,
ironically, leave you wondering "why am I filling out this template?"

Some specific things I've noticed:

1. It doesn't provide any context as to why each section exists
2. Some sections are optional, yet often people feel obliged to fill
   them out. This will inevitably feel like busy-work and degrade the
   experience of the PR author as well as the PR reader.
3. If someone wants to educate themselves about writing good pull
   request, it doesn't actually provide them with any resources to do
   so.

This expands unpacks each section of the pull request and gives a
purpose and approach for each section. In this way, it actually
mirrors the structure of the PR itself in a small way.

All of the section explanations have been moved into
comments. The main reason is so that you don't have to delete them
from your PR description, and can refer back to them as you develop
it. It's a lot of information, and so this ought to give folks the
opportunity to learn as they go and also to refine within the same
pull request.

Wherever it makes sense, I'v tried to link appropriate tools and
resources to learn about and build good PRs.

Finally, I've marked optional seccions as being optional to cut down
on the feeling of busy-work.