Persist Relationship Data in Store
As we navigate through the eholdings app, data from resolved requests
is persisted in the redux store.  When we're loading a search route
with a selected detail record, one request is fired off for the list
of search records and one is fired off for the singleton detail
record.  If the search request resolves after the detail request, the
detail page's related records are overwritten by the "newer" request,
which contains no relationship data.  The result in the UI is the list
of related records on a detail page flickering and disappearing.

This fix introduces a helper method to intelligently merge
relationship data.  Once a relationship is established in the store,
only updates to that data are accepted.  If a newer request returns
with no relationship data whatsoever, the older data is preferred.