Preserve search and filter options when closing preview pane
In the three-pane layout, the preview pane can be hidden via the 'x'
icon in the pane header.  This was previously a hard link to
`/eholdings`, which would blow away the search results as well as the
preview pane.  This work uses `router.history.goBack()` instead of a
hard link.  Also added some tests specific to the 'x' button, as our
previous tests only exercised the '<-' button on a full detail record.