Make toggle switch accessible
Initial the input was set to display none as such when trying
to navigate the form via keyboard the toggle switch / input
would be skipped as it was hidden.  Here we move the input
behind the style slider which allows the input to accessible
in and still usable to via press or click.