Change modal message for custom-title deselection
Deselecting a custom-title is a destructive operation
and if its the last instance of the custom-title there is no
way to then get it back so there needs to be messaging
for the user.

We have access to the to title.resources if the user navigates
to the resource from a search. However, one caveat if the URL
is pasted into the URL this title.resources Array is empty.