Add check for inFlight isSelect
Only show sections only if isSelect is not inFlight.
Various sections in the View would show up before
the isSelect would resolve and allow users to
interact with those areas if they were quick
enough to do so before the resolution of the selection.
If they were so quick, whatever they had on the
screen would disspear as the resolution took place.
We want to be defensive and explicit about when
certain areas should show up.