UIEH-406 - Remove Holding Status Toggle Switch in Favor of Dropdown / Button
The "Select / Deselect" toggle switch was increasingly out of place in
the current UI so we're abandoning it in favor of an action in the
pane header.  Additionally we're adding a button inline with the field
if the package is not in the librarian's holdings.  This latter piece
should eliminate any ambiguity if a new user is approaching a package
and unsure how to select it.