Extract selection status into its own component.
We're going to be changing this interaction and it's difficult to see
how it interacts with its enclosing environment. In order to work more
surgically on it we want to disentagle it from its parent scope.

We aren't changing anything about it, just putting gates around the
data that flows into it. Even the internal variables didn't change at
all and we add aliases to the props.