Switch to a simple 'addPackageToHoldings' action.
The toggle UI is gone, but the action handlers are still oriented
around toggles. However, this complicated things by having to deal
with both cases "selected vs unselected" all the way down the
stack. This decition is know at the very moment we click the "add to
holdings" button or "remove from holdings", so our action handlers do
not need to account for both cases.

This creates a single purpose action handler for adding a package and
binds it to both the dropdown menue and the button on the detail
record.

Later on, we can factor out the "remove from holdings" action.