Lift pending state from the view into the model.
Our application concentrates state and the actions that
change it within the route, and then the representations of that state
go into the views. As a 99/1 rule, our views should be pure functions.

The PackageShow component maintains a lot of state internally, that
belongs "higher up".

This move a piece of that state (whether or not a package is in
filght) up to the model layer, thereby simplifying the APIs of the
components that get rendered down the stack.