Capability to block and unblock mirage server in test suite.
Sometimes when you are writing acceptance tests, you need to assert on
some sort of intermediate state. Usually this is something that
represents a in-flight server request like a loading spinner or
otherwise. Testing this can be a challenge, because you are dependent
on the timing of when the request returns from the server, and you
cannot be fully in control from your test case about when to make
assertions and when not to.

What is needed is a way to _explicity_ control request timing.

This adds a `block()` method to the mirage server which will cause all
requests to not return indefinitely until the `unblock()` method is
invoked. Now, instead of tweaking timing of server responses, you can
just block() inside a `describe.beforeEach`, and be guaranteed that all of the
correspending it blocks inside that describe block will be run while
the server is in a pending state.

before:

```js

describe('some asynchronous operation', () => {
  beforeEach(function () {
    this.server.timing = 50; // all request are delayed by 50ms.
    return interator.clickTheThing();
  })
  afterEach(function () {
    this.server.timing = 0;
  })
  it('is working', () => {
    expect(interactor.isWorking).to.equal(true);
  })
  describe('when the operation returns', () => {
    it('shows this thing', () => {
      expect(interator.thisThing.isVisible).to.equal(true);
    });
  })
})

```

after:

```js
describe('some asynchronous operation', () => {
  beforeEach(function () {
    this.server.block(); // all requests are blocked now indefinitely.
    return interactor.clickTheThing();
  })
  it('is working', () => {
    expect(interactor.isWorking).to.equal(true);
  })
  describe('when the operation returns', () => {
    beforeEach(function () {
      this.server.unblock(); // all blocked operations, now return.
    })
    it('is no longer working', () => {
      expect(interactor.isWorking).to.equal(false);
    })
    it('shows this thing', () => {
      expect(interactor.thisThing.isVisible).to.equal(true);
    })
  })
})
```

Note that this is only needed when you want to test the _sequence_ of
intermediate states. The above hypothetical test case would work with
BigTest without touching the server timing at all. Consider:

```js
describe('some asynchronous operation', () => {
  beforeEach(function () {
    return interactor.clickTheThing();
  })
  it('is working', () => {
    expect(interactor.isWorking).to.equal(true);
  })
  it('is no longer working', () => {
    expect(interactor.isWorking).to.equal(false);
  })
  it('shows this thing', () => {
    expect(interactor.thisThing.isVisible).to.equal(true);
  })
});

```

Because of the way BigTest convergent assertions work, it will ensure
that the loading state is rendered, and also that it is not
rendered. As long as both of those conditions happen within the 2
second window, the test will pass. What it _doesn't_ ensure is
that *first* the loading state happens and *then* that it doesn't
happen.