[UIEH-499] Upgrade to React 16.4.x (#496)
* Upgrade react

* Upgrade react-dom too

* WIP: lift state out of `search-form` component

This is because `getDerivedStateFromProps` now gets called on:
- Forced update
- setState
- props changed

And there's no way to tell which should be the source of truth (the
props or state? Which changed?). In order to track that we would have
to store props in state, which seems gross.

The best option seemed to remove state from this component entirely
and just have it take props.

* Unprettify file

* Passing tests! Lifted state entirely out of the search-form component

* Rename callback prop names for `search-form`

* Fix rebase confict & fix tests

* Shape `searchFilter` prop.