UIEH-757 restore eHoldings to settings
stripes-core uses `settings.${module}.enabled` to determine whether a
module should be listed on the settings page. Unfortunately, it was
removed in #840 because I didn't realize that requirement.

Here, it's being restored BUT without `visible: true` because we do not
want to grant this permission to anybody; we want granular, per-item
permissions as UITEN-35 indicates.

Fixes UIEH-757; refs UITEN-35